refactor: refactor the project and enhance the openfile at line
